Switching to std::variant for VS2017-15.5
- update mpark::variant to newest version
