+++ cmake/Phylanx_Utils.cmake
@@ -20,6 +20,7 @@ phylanx_include(
   AddTest
   AddPythonTest
   CreateSymbolicLink
+  DetectCppDialect
   ExportTargets
   ForceOutOfTreeBuild
   GetPythonExtensionLocation
