+++ phylanx/util/variant.hpp
@@ -8,10 +8,24 @@
 #define PHYLANX_UTIL_VARIANT_HPP
 
 #include <phylanx/config.hpp>
+
+#if defined(_MSC_VER) && _MSC_VER >= 1912
+#include <variant>
+#else
 #include <phylanx/util/detail/variant.hpp>
+#endif
 
 namespace phylanx { namespace util
 {
+#if defined(_MSC_VER) && _MSC_VER >= 1912
+    using std::variant;
+    using std::monostate;
+
+    using std::holds_alternative;
+    using std::get;
+    using std::get_if;
+    using std::visit;
+#else
     using mpark::variant;
     using mpark::monostate;
 
@@ -19,6 +33,7 @@ namespace phylanx { namespace util
     using mpark::get;
     using mpark::get_if;
     using mpark::visit;
+#endif
 
     ///////////////////////////////////////////////////////////////////////////
     // class template recursive_wrapper
