+++ src/CMakeLists.txt
@@ -107,6 +107,13 @@ set_property(TARGET phylanx_component APPEND
 
 set(phylanx_targets ${phylanx_targets} phylanx_component)
 
+get_property(PHYLANX_TARGET_COMPILE_OPTIONS_VAR
+  GLOBAL PROPERTY PHYLANX_TARGET_COMPILE_OPTIONS)
+foreach(_flag ${PHYLANX_TARGET_COMPILE_OPTIONS_VAR})
+  target_compile_options(phylanx_component INTERFACE ${_flag})
+  target_compile_options(phylanx_component PUBLIC ${_flag})
+endforeach()
+
 install(
   TARGETS ${phylanx_targets}
   EXPORT PhylanxTargets
